chore(stucture): split dtos into separate directories
