+++ FastFoodOperator.Api/DTOs/Category/CategoryResponseDto.cs
@@ -1,4 +1,4 @@
-namespace FastFoodOperator.Api.DTOs;
+namespace FastFoodOperator.Api.DTOs.Category;
 
 public class CategoryResponseDto
 {
