+++ FastFoodOperator.Api/DTOs/Category/CategoryUpdateDto.cs
@@ -1,4 +1,4 @@
-namespace FastFoodOperator.Api.DTOs;
+namespace FastFoodOperator.Api.DTOs.Category;
 
 public class CategoryUpdateDto
 {
