+++ FastFoodOperator.Api/DTOs/Combo/ComboProductsUpdateDto.cs
@@ -1,4 +1,4 @@
-namespace FastFoodOperator.Api.DTOs;
+namespace FastFoodOperator.Api.DTOs.Combo;
 
 public class ComboProductsUpdateDto
 {
