+++ FastFoodOperator.Api/DTOs/Combo/ComboResponseDto.cs
@@ -1,4 +1,4 @@
-namespace FastFoodOperator.Api.DTOs;
+namespace FastFoodOperator.Api.DTOs.Combo;
 
 public class ComboResponseDto
 {
