+++ FastFoodOperator.Api/DTOs/Product/ProductIngredientUpdateDto.cs
@@ -1,4 +1,4 @@
-namespace FastFoodOperator.Api.DTOs;
+namespace FastFoodOperator.Api.DTOs.Product;
 
 public class ProductIngredientUpdateDto
 {
